backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["rxmi-bkd <boukadidarami0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
drf-standardized-errors = {extras = ["openapi"], version = "^0.14.1"}


[tool.poetry.group.dev.dependencies]
django-cors-headers = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
